Contracts formatted using prettier

+++ test/RewardControlHarnessV2.sol
@@ -3,14 +3,13 @@ pragma solidity ^0.4.24;
 import "./RewardControlHarness.sol";
 
 contract RewardControlHarnessV2 is RewardControlHarness {
-    uint value;
+    uint256 value;
 
     function getValue() public view returns (uint256) {
         return value;
     }
 
-    function setValue(uint _value) public {
+    function setValue(uint256 _value) public {
         value = _value;
     }
-
 }
